Move format to common module
