--- indexer/src/main/java/io/dongyue/at15/tree/indexer/format/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Created by at15 on 15-12-26.
- */
-package io.dongyue.at15.tree.indexer.format;
