+++ indexer/src/main/java/io/dongyue/at15/tree/indexer/mapreduce/meta/MetaMapper.java
@@ -1,6 +1,6 @@
 package io.dongyue.at15.tree.indexer.mapreduce.meta;
 
-import io.dongyue.at15.tree.indexer.format.Meta;
+import io.dongyue.at15.tree.common.format.Meta;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
