+++ indexer/src/main/java/io/dongyue/at15/tree/indexer/mapreduce/sort/SortReducer.java
@@ -1,6 +1,6 @@
 package io.dongyue.at15.tree.indexer.mapreduce.sort;
 
-import io.dongyue.at15.tree.indexer.format.Meta;
+import io.dongyue.at15.tree.common.format.Meta;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
